requirements.txt

click==8.1.7
docker==7.0.0
requests>=2.31.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
pydantic==2.5.0
typer>=0.9.0
rich==13.7.0
PyYAML==6.0.1
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.8.0
isort==5.13.2